Implement support for EVP_chacha20_poly1305
ChaCha20-Poly1305 is only exposed via the BoringSSL EVP_AEAD interface. While it provides a cleaner API to our customers it lacks compatibility with the corpus of applications that are developed against upstream OpenSSL's EVP_CIPHER interface. This PR implements that interface and aims for compatibility with existing consumers of that interface.

From a behavioral perspective, the fundamental difference between the EVP_AEAD and EVP_CIPHER interface is that they operate as one-shot and streaming respectively. By providing the EVP_CIPHER interface, we also introduce ChaCha20-Poly1305 streaming (i.e. the ability to run the the cipher in arbitray increments down to a single byte) support as a new capability.

We implement this by leveraging a fallback ChaCha20-Poly1305 implementation that already exists in AWS-LC which traces back to upstream OpenSSL and supports streaming. This implementation is still assembly optimized, but presumably not as performant as the preferred implementation. We then draw inspiration from upstream OpenSSL to implement the EVP_CIPHER APIs.

For testing, we use the existing EVP_AEAD test vectors and run them through the EVP_CIPHER test harness to maintain coverage parity between the implementations.