Remove X509_VERIFY_PARAM_get0_peername
This allowed the caller to query which DNS SAN actually matched. It's
not an unreasonable API, though putting it on X509_VERIFY_PARAM is a
little odd, as is not having it for other name types. It just wasn't
const-correct. I went to fix that and found no one was using it.

We can put it back if we need it. Most of the underlying machinery needs
to stay anyway because X509_check_host has an output parameter and Node
exposes it. (Unclear if Node actually has a use case. I suspect they
just returned it because it was available.)

Change-Id: Id7f15016e29189bc4f3cebb33c680279c8cb3178
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/64252
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: David Benjamin <davidben@google.com>
(cherry picked from commit f78fe19fc98e0e6f760e05c6b9d48725004700d0)