Read fips scope marker addresses using adrp instead of adr to increase reach (#1332)
When including the new Ed25519 implementations in the FIPS build the follow error occurred:
bcm.c:47662: Error: pc-relative address offset out of range
The reason is that we get the address of e.g. BORINGSSL_bcm_text_end using adr like this:
adr x4, BORINGSSL_bcm_text_end
But since the size of the fips module has slowly increased the last few years the amount of memory we attempt to jump with adr runs over the reach of the instruction (about 1MB or so).

The solution is to no longer consider BORINGSSL_bcm_text_[end,start] for special and use a redirector function to resolve their addresses.