Print running times of proofs, factor out proof runner
This patch makes `*.correct` files contain running time in seconds as
well as its start and end time.
This is to find out which proofs take most of the time from CI.
According to the CPU usage graph, there are only two or three proofs
that use most of the CI running time, and they even seem to start late during
parallel runs. If they start early, we can reduce significant amount of
CI time!
As a part of this update, this patch factors out proof runner script commands from inline commands in
Makefile as a separate shell script file.