Cipher-stealing for 1 trailing block.
Test vector added for 2 blocks + x bytes.
The difference of 1 trailing block between the test vector
and the exit point is due to that at the beginning of the re-encrypt
function, one block is deducted from the length.
That's because it follows the structure of xts_decryt which accounts for
1 full block and 1 partial block in cipher-stealing handling.