Fix decoding of MOVSXD in the x86 model
This corrects the order of sizes in the non-trivial 32->64 case,
as well as properly distinguishing and handling the (admittedly
degenerate) 32->32 case. The other degenerate case 16->16 is
rejected by design, as with most other instructions using operand
size override prefixes.

This also adds new cases to the simulator to validate MOVZX and
MOVSX/MOVSXD, which are not currently used in s2n-bignum code and
hence were not previously cosimulated against hardware in standard
checks.