Specify the behavior of bignum_inv_p25519 in degenerate cases
Instead of having the postcondition guarded by an assumption of
coprimality, there is now a precise case analysis in the specification
asserting that it returns zero in the degenerate cases. For now, this
is worked around in other proofs, but it gives scope for actually
avoiding the special-case reasoning about zero in many such
applications.