Tidy up Montgomery ladders
This simplifies the remaining Montgomery ladder implementations by
avoiding the special code for zero handling, since it is not actually
necessary given the behavior of the modular inverse in this case. In
addition, the proofs have been tidied up a bit, factoring out the
basic mathematics so that the loop invariant becomes simpler.

The update also fixes a README typo pointed out by Dan Bernstein
and removes a couple of stray comments arising from SLOTHY output
in the Lenngren-derived X25519 code.