Modify specification style for Weierstrass curve point operations
The previous specifications proved that, assuming none of the usual
degenerate cases, the results correspond to "canonical" Jacobian
operations, with specific encoding of the point at infinity as
(1,1,0). However, this canonical form seems to have little or no
implementation value and any interesting conclusions are excluded by
the exceptions in the precondition.

The new specifications are stated in terms of the mapping back to
affine form, for which any triple of the form (x,y,0) is treated as
the point at infinity. This already removes all special cases from
doubling and removes one of the four special cases from addition,
namely x + -x' = 0 where x' and x are projectively equivalent. The
new specifications are also perhaps a bit more compact and easily
understood.