Remove one redundant instruction, tidy a few proofs with WORD_BLAST
There was a pointless duplication of the first instruction of
the x86 version of bignum_mul_p521_alt, moving one of the input
arguments out of rdx.

A number of explicit incantations expanding sums bitwise are
now replaced by simple applications of WORD_BLAST, which is
in effect automating the same sort of pattern, and more.