Only enable dilithium and secp256k1 benchmark if AWS-LC API supports it (#1495)
### Description of changes: 
Handle newer versions of speed.cc being built with older versions of
AWS-LC that don't have dilithium or secp256k1. This is the same check we
have [down
below](https://github.com/aws/aws-lc/blob/main/tool/speed.cc#L2718) but
forgot this spot.

### Call-outs:
This is helpful for the canary.

### Testing:
Trying to test locally but running into compiler/delocate issues. The
new check in the CI should handle this all.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.