Bump mysql integration CI to 8.3 (#1508)
MySQL 8.2/8.3 changed some of the build behavior with OpenSSL. According
to the documentation, `-DWITH_SSL={path_name}` is a viable option, but
it only seems to work with Visual Studio and MacOS due to the usage of 
`${CMAKE_CFG_INTDIR}`. `${CMAKE_CFG_INTDIR}` is undefined on single
build configs like regular Make and Ninja, so our build fails when using the
option.

Another option was documented in mysql:
> Another permitted way to do the same thing is to set WITH_SSL to
system and set the CMAKE_PREFIX_PATH option to path_name.)

Using this gets around our CI build issues with mysql 8.3.