Support vpinsrq in delocater (#1543)
Instead of generating portable code, gcc can be configured with e.g. march=cpu-type that allows it to generate code using instructions from instruction sets supported on up-to cpu-type.

This fails because vpinsrq is not supported in the delocater.

This PR adds support for vpinsrq an instruction from the set AVX512DQ. I don't think there are other 4-argument instructions, for where a GOT reloc can be emitted, that we need to support right now. This also makes the implementation a tad easier, because we do not need to cater for the relocation being either the first or second argument - it can only be the second.

Otherwise, the implementation follows the one for the type instrThreeArg.

(cherry picked from commit d9401623bf21394de5d34e3b49c506fb7af235b5)