Fix skipped tests in Mariadb integration CI (#1533)
1. Apparently the new errors we recently turned off are due to the
default auto-chaining behavior in OpenSSL. I applied a patch to enable
auto-chaining in Mariadb, since consumers will likely be looking for the
default behavior Mariadb tries to enable.
2. Not sure what desc_index_min_max does, but it does seem to have a
history of being flaky based on this commit: MariaDB/server@03fa2c3.
It seems unrelated to AWS-LC, so I think it's alright to leave off.
3. My guess is that ssl_crl is trying to connect to the server containing
the crl, but there are instances where the crl server is unresponsive
(possibly due to too much load on the server, common issue with crls).
Why our CI still fails, is due the error message being overwritten in the
original patch. The patch correctly encapsulates new AWS-LC errors,
but we've also overwritten the potential error message that was being
expected. I've generated a new patch for the CI.