Fix ec2 CI testing framework (#1541)
This fixing ec2 CI testing framework we've turned off. I tried
introducing intentional test failures to see if the framework will fail
correctly (as shown in the commit history of this PR). I also introduced
10 builds to run simultaneously, to see if we would get the sporadic
test failures we had been having.
The general run fails early in the build and the FIPS specific one fails
sooner since it's being ran after the sanitizer tests. Good news is all
runs fail and succeed as anticipated without the original sporadic
issues we've been having.

1. Apparently the docker container needs to be in "privileged-mode" for
the TSAN tests to work, so I've turned that on in the SSM document.
2. An sporadic error message `E: Unable to lock the administration
directory (/var/lib/dpkg/) is another process using it? ` occurs when we
try to call `apt-get update` right after the instance is spun up.
Calling `killall apt apt-get` effectively gets around the issue.

More details on debugging can be found in thread of `P113131493`