Bifurcate microarchitectural variants of Weierstrass point additions
As with the earlier update for doublings, the Jacobian point adidtion
and mixed addition operations for the curves P-256, P-384, P-521,
secp256k1 and SM2 now all have the usual two versions targeting
different microarchitectures, one of them called "_alt", following the
general s2n-bignum convention.

The "_alt" forms for ARM now present are just renamed versions of the
originals (which were based on "_alt" field operations), with the new
code taking over the old non-alt name. For x86 the non-alt ones are
the same as before and the "_alt" forms are new.