Fix MySQL version tag alarm (#1585)
We got a false alarm when we updated the mysql version tag strategy.
Apparently my ubuntu box and the docker image had different versions of
git and this caused a behavioral difference when running `git tag
--sort=-taggerdate`.

The docker image has a newer version (`2.34.1`) and gives the tag date
in ascending order. This caused us to compare against the oldest version
of mysql instead and triggers the alarm. This PR fixes that.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.