[Backport] Prevent non-constant-time code in Kyber-R3 implementation
A branch can be introduced by the compiler from the `poly_frommsg` code.
This has been reported for Clang versions
15-18, x86_64, for the following compiler flags:
- `-Os`
- `-O1`
- `-O2 -fno-vectorize`
- `-O3 -fno-vectorize`

An upstream fix has been published
[here](https://github.com/pq-crystals/kyber/commit/9b8d30698a3e7449aeb34e62339d4176f11e3c6c)
which we are deciding not to take directly. Due to the way AWS-LC
compiles the Kyber reference code, it may not have the
desired effect. We instead opt to use our own constant-time
implementations for performing conditional bit-masks which
have much stronger guarantees around portability.

[Commit in main](https://github.com/aws/aws-lc/commit/4b07805bddc55f68e5ce8c42f215da51c7a4e099]
note: In main, the fix was also applied to ml-kem-ipd which does not
exist in this branch.