Fix AES key size for AES256 in ABI test (#1629)
### Issues:
Resolves V1401987253

### Description of changes: 
This key buffer is too small for the larger key sizes that the test
uses.

### Call-outs:
It might be worth adding a address sanitizer + ABI test dimension. So
far we haven't hit this before so I didn't add a new docker image with
everything installed for that.

### Testing:
Ran locally and manually reviewed the change for memory safety. 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.