[Backport] Prevent non-constant-time code in Kyber-R3 implementation (#1632)
### Issues:
Resolves V1399146249

### Description of changes: 
Backports PR #1619. This change is isolated to the Kyber-round3
implementation
which is outside of the FIPS boundary.

### Call-outs:
The original PR applied the mitigation to both Kyber-r3 and ML-KEM-IPD
implementations. Only the former exists in this branch.

### Testing:
How is this change tested (unit tests, fuzz tests, etc.)? Are there any
testing steps to be verified by the reviewer?

CI

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.