align gcc version with curl's CI (#1633)
Curl's tip of main is failing against a compiler warning specific to
gcc-12 at the moment with `-Werror` enabled. This is causing errors in
our integration CI. We enable the flag since Curl's CI enables it, but
they run against gcc-11 for their
CI at the moment.

I've submitted an issue to curl to fix it
(https://github.com/curl/curl/issues/13932), but we should use gcc-11
for our Curl CI to avoid with curl's main branch that are out of our
control.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.