Add EC seed functions as deprecated no-ops (#1674)
Ruby exposes the `EC_GROUP` seed functions and allows you to set the
seed with `EC_GROUP_set_seed`. It's one thing that are `EC_GROUP`s are
static and immutable, but setting the seed is more prevalent to custom
curves. We don't encourage using custom curves, so I've chose to mark it
as a no-op and deprecated.
We could arguably support `EC_GROUP_get0_seed` and
`EC_GROUP_get_seed_len` for our named curves, but the seed value is only
used during the initial curve parameter generation. We've chosen to mark
this as a no-op for now, implementing them creates additional complexity
that doesn't really provide additional value to the consumer.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.