add support for EC_POINT_bn2point (#1645)
Ruby happens to consume this.
This is deprecated in OpenSSL 3.0 and I've marked it likewise in AWS-LC
as well. Although Ruby has generally moved off of deprecated 3.0 APIs,
they still depend on this and later versions like 3.2 and 3.3 have not
removed support for this either.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.