Set ret to NULL before return in EC_POINT_bn2point (#1692)
Static analysis was complaining about us using the pointer after it was
freed. I don't believe this is an actual "issue", but we might as well
make the NULL explicit before returning.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.