add ECPKParameters_print as no-op (#1686)
Ruby consumes  `ECPKParameters_print`. The entire function is a bit
convoluted and not really beneficial unless you're using custom curves.
We can expose this as a no-op until there's a more proper use case for
this.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.