Avoid using term_match for performance, do small refactoring to equiv tactics
This patch makes small updates to avoid term_match which seems to be slower than directly using the match statement.
A particular case that had to be fixed was when term_match was being invoked for every assumption because the number
of assumptions can grow linearly to the number of simulated instructions unless it is cleaned up in the middle.
This fix specifically brought ~14% speedups to proofs using equivalence checking tactic because when the lockstep simulation
simulates one program it had to filter assumptions to hide that correspond to another program and it was
using term_match.
There was also some speedups in large proofs not using equivalence tactic. I will attach the full result in the PR.

Another update that was orthogonally make was to the code related to equivalence checks :

- Rename `mk_equiv_statement` to `mk_equiv_statement_simple` that only considers a pair of straight-line codes from begin to end,
and add `mk_equiv_statement` for more generic case.
- Some fix in `common/relational2.ml` in order to support constant-time proof in the future.