Make subroutine simulation tactic robust for write-free subroutines
Previously the respective SUBROUTINE_SIM_TAC implementations were
assuming that the subroutine correctness theorem had some additional
conditions as the antecedent of an implication in its conclusion
(usually around nonaliasing). Now they work even if this is not the
case, e.g. if the core subroutine does not write memory or have other
special preconditions.