Add P-256 precomputed point scalar multiplication operation
Given a scalar n and a precomputed table of multiples of a point P on
the P-256 curve, this produces the scalar product n * P. The assumed
size of the table of multiples is determined by an additional
"blocksize" parameter determining how many bits of scalar are handled
per table block. The formal proof assumes that this is in the range
2 <= blocksize <= 31, but since the table size grows exponentially
with blocksize, values over 10 are implausible anyway. The benchmark
program just exercises sizes 4--6 and testing 2--9. See the top
comment banner in the code for more on the precomputed table format.